backend/requirements.txt
flask==3.1.*
flask-cors==5.0.*
mysql-connector-python==9.0.*
pytest==8.3.*
